feat: Separar históricos
