+++ tsconfig.app.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/app.tsx","./src/env.ts","./src/main.tsx","./src/routes.tsx","./src/vite-env.d.ts","./src/api/get-anamnese.ts","./src/api/get-exam-schedule.ts","./src/api/get-exams.ts","./src/api/get-schedule.ts","./src/api/get-specialists.ts","./src/api/get-specialties.ts","./src/api/get-user.ts","./src/api/get-users.ts","./src/api/register-exam-schedule.ts","./src/api/register-patient.ts","./src/api/register-schedule.ts","./src/api/sign-in.ts","./src/api/sign-out.ts","./src/api/sign-up.ts","./src/api/update-profile.ts","./src/components/account-menu.tsx","./src/components/feature-card.spec.tsx","./src/components/feature-card.tsx","./src/components/header.spec.tsx","./src/components/header.tsx","./src/components/historic-table.tsx","./src/components/new-schedule.tsx","./src/components/times-slots.spec.tsx","./src/components/times-slots.tsx","./src/components/user-profile-sheet.tsx","./src/components/ui/alert-dialog.tsx","./src/components/ui/button.tsx","./src/components/ui/calendar.tsx","./src/components/ui/card.tsx","./src/components/ui/command.tsx","./src/components/ui/dialog.tsx","./src/components/ui/dropdown-menu.tsx","./src/components/ui/input.tsx","./src/components/ui/label.tsx","./src/components/ui/password-input.tsx","./src/components/ui/popover.tsx","./src/components/ui/separator.tsx","./src/components/ui/sheet.tsx","./src/components/ui/skeleton.tsx","./src/components/ui/table.tsx","./src/components/ui/tabs.tsx","./src/components/ui/toast.tsx","./src/components/ui/toaster.tsx","./src/components/ui/toggle-group.tsx","./src/components/ui/toggle.tsx","./src/components/ui/use-toast.ts","./src/lib/axios.ts","./src/lib/react-query.ts","./src/lib/utils.ts","./src/pages/404.tsx","./src/pages/error.tsx","./src/pages/_layouts/app.tsx","./src/pages/_layouts/auth.tsx","./src/pages/app/historic.tsx","./src/pages/app/home.tsx","./src/pages/app/schedules.tsx","./src/pages/auth/sign-in.tsx","./src/pages/auth/sign-up.tsx","./src/utils/axiosErrorHandler.ts"],"version":"5.6.3"}
+{"root":["./src/app.tsx","./src/env.ts","./src/main.tsx","./src/routes.tsx","./src/vite-env.d.ts","./src/api/get-anamnese.ts","./src/api/get-exam-schedule.ts","./src/api/get-exams.ts","./src/api/get-schedule.ts","./src/api/get-specialists.ts","./src/api/get-specialties.ts","./src/api/get-user.ts","./src/api/get-users.ts","./src/api/register-exam-schedule.ts","./src/api/register-patient.ts","./src/api/register-schedule.ts","./src/api/sign-in.ts","./src/api/sign-out.ts","./src/api/sign-up.ts","./src/api/update-profile.ts","./src/components/account-menu.tsx","./src/components/feature-card.spec.tsx","./src/components/feature-card.tsx","./src/components/header.spec.tsx","./src/components/header.tsx","./src/components/historic-table.tsx","./src/components/new-schedule.tsx","./src/components/times-slots.spec.tsx","./src/components/times-slots.tsx","./src/components/user-profile-sheet.tsx","./src/components/ui/alert-dialog.tsx","./src/components/ui/button.tsx","./src/components/ui/calendar.tsx","./src/components/ui/card.tsx","./src/components/ui/command.tsx","./src/components/ui/dialog.tsx","./src/components/ui/dropdown-menu.tsx","./src/components/ui/input.tsx","./src/components/ui/label.tsx","./src/components/ui/password-input.tsx","./src/components/ui/popover.tsx","./src/components/ui/separator.tsx","./src/components/ui/sheet.tsx","./src/components/ui/skeleton.tsx","./src/components/ui/table.tsx","./src/components/ui/tabs.tsx","./src/components/ui/toast.tsx","./src/components/ui/toaster.tsx","./src/components/ui/toggle-group.tsx","./src/components/ui/toggle.tsx","./src/components/ui/use-toast.ts","./src/lib/axios.ts","./src/lib/react-query.ts","./src/lib/utils.ts","./src/pages/404.tsx","./src/pages/error.tsx","./src/pages/_layouts/app.tsx","./src/pages/_layouts/auth.tsx","./src/pages/app/historic-anamnesis.tsx","./src/pages/app/historic-exam-schedule.tsx","./src/pages/app/historic-schedule.tsx","./src/pages/app/home.tsx","./src/pages/app/schedules.tsx","./src/pages/auth/sign-in.tsx","./src/pages/auth/sign-up.tsx","./src/utils/axiosErrorHandler.ts"],"version":"5.6.3"}
